+++ packages/app-utils/src/components/index.ts
@@ -5,8 +5,5 @@ export * from './FontBootstrap';
 export * from './FontsLoaded';
 export * from './PluginsBootstrap';
 export * from './ThemeBootstrap';
-export * from './usePlugins';
-export * from './useConnection';
 export * from './useServerConfig';
 export * from './useUser';
-export * from './useLoadTablePlugin';