+++ packages/code-studio/src/main/AppInit.tsx
@@ -18,6 +18,7 @@ import {
 import { FileStorage } from '@deephaven/file-explorer';
 import { useApi, useClient } from '@deephaven/jsapi-bootstrap';
 import type { dh as DhType, IdeConnection } from '@deephaven/jsapi-types';
+import { useConnection } from '@deephaven/jsapi-components';
 import {
   DecimalColumnFormatter,
   getSessionDetails,
@@ -44,14 +45,9 @@ import {
   Workspace,
   WorkspaceStorage,
   ServerConfigValues,
-  DeephavenPluginModuleMap,
 } from '@deephaven/redux';
-import {
-  useConnection,
-  usePlugins,
-  useServerConfig,
-  useUser,
-} from '@deephaven/app-utils';
+import { useServerConfig, useUser } from '@deephaven/app-utils';
+import { type PluginModuleMap, usePlugins } from '@deephaven/plugin';
 import { setLayoutStorage as setLayoutStorageAction } from '../redux/actions';
 import App from './App';
 import LocalWorkspaceStorage from '../storage/LocalWorkspaceStorage';
@@ -76,7 +72,7 @@ interface AppInitProps {
   setLayoutStorage: (layoutStorage: LayoutStorage) => void;
   setDashboardConnection: (id: string, connection: IdeConnection) => void;
   setDashboardSessionWrapper: (id: string, wrapper: SessionWrapper) => void;
-  setPlugins: (map: DeephavenPluginModuleMap) => void;
+  setPlugins: (map: PluginModuleMap) => void;
   setUser: (user: User) => void;
   setWorkspace: (workspace: Workspace) => void;
   setWorkspaceStorage: (workspaceStorage: WorkspaceStorage) => void;