+++ packages/embed-grid/src/App.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useMemo, useState } from 'react';
-import { useConnection, useUser } from '@deephaven/app-utils';
+import { useUser } from '@deephaven/app-utils';
+import { useConnection } from '@deephaven/jsapi-components';
 import { ContextMenuRoot, LoadingOverlay } from '@deephaven/components'; // Use the loading spinner from the Deephaven components package
 import {
   InputFilter,