+++ packages/jsapi-components/src/index.ts
@@ -1,3 +1,4 @@
+export * from './ConnectionContext';
 export * from './HookTestUtils';
 export { default as TableInput } from './TableInput';
 export * from './RefreshTokenBootstrap';
@@ -5,6 +6,7 @@ export * from './RefreshTokenUtils';
 export { default as useBroadcastChannel } from './useBroadcastChannel';
 export { default as useBroadcastLoginListener } from './useBroadcastLoginListener';
 export * from './useCheckIfExistsValue';
+export * from './useConnection';
 export { default as useDebouncedViewportSearch } from './useDebouncedViewportSearch';
 export * from './useDebouncedViewportSelectionFilter';
 export * from './useFilterConditionFactories';